+++ tensorflow/python/estimator/BUILD
@@ -37,7 +37,6 @@ py_library(
         ":parsing_utils",
         ":run_config",
         ":training",
-        ":warm_starting_util",
         "//tensorflow/python:util",
     ],
 )
@@ -278,12 +277,12 @@ py_library(
     srcs = ["canned/dnn_testing_utils.py"],
     srcs_version = "PY2AND3",
     deps = [
+        ":estimator",
         ":head",
         ":metric_keys",
         ":model_fn",
         ":numpy_io",
         ":prediction_keys",
-        ":warm_starting_util",
         "//tensorflow/core:protos_all_py",
         "//tensorflow/python:array_ops",
         "//tensorflow/python:check_ops",
@@ -427,7 +426,6 @@ py_library(
         ":model_fn",
         ":run_config",
         ":util",
-        ":warm_starting_util",
         "//tensorflow/core:protos_all_py",
         "//tensorflow/python:client",
         "//tensorflow/python:control_flow_ops",
@@ -866,41 +864,5 @@ py_test(
         "//tensorflow/python:framework_ops",
         "//tensorflow/python:session",
         "//tensorflow/python:training",
-    ],
-)
-
-py_library(
-    name = "warm_starting_util",
-    srcs = ["warm_starting_util.py"],
-    srcs_version = "PY2AND3",
-    deps = [
-        "//tensorflow/python:array_ops",
-        "//tensorflow/python:framework_ops",
-        "//tensorflow/python:platform",
-        "//tensorflow/python:state_ops",
-        "//tensorflow/python:training",
-        "//tensorflow/python:variable_scope",
-        "//tensorflow/python:variables",
-        "//tensorflow/python/feature_column",
-    ],
-)
-
-py_test(
-    name = "warm_starting_util_test",
-    size = "small",
-    srcs = ["warm_starting_util_test.py"],
-    srcs_version = "PY2AND3",
-    deps = [
-        ":warm_starting_util",
-        "//tensorflow/python:array_ops",
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python:dtypes",
-        "//tensorflow/python:framework_ops",
-        "//tensorflow/python:init_ops",
-        "//tensorflow/python:training",
-        "//tensorflow/python:variable_scope",
-        "//tensorflow/python:variables",
-        "//tensorflow/python/feature_column",
-        "//third_party/py/numpy",
     ],
 )
