+++ tensorflow/python/training/training.py
@@ -95,6 +95,8 @@ See the @{$python/train} guide.
 @@load_variable
 @@list_variables
 @@init_from_checkpoint
+@@warm_start
+@@VocabInfo
 """
 
 # Optimizers.
@@ -188,6 +190,8 @@ from tensorflow.python.training.training_util import get_global_step
 from tensorflow.python.training.training_util import assert_global_step
 from tensorflow.python.training.training_util import create_global_step
 from tensorflow.python.training.training_util import get_or_create_global_step
+from tensorflow.python.training.warm_starting_util import VocabInfo
+from tensorflow.python.training.warm_starting_util import warm_start
 from tensorflow.python.pywrap_tensorflow import do_quantize_training_on_graphdef
 from tensorflow.python.pywrap_tensorflow import NewCheckpointReader
 from tensorflow.python.util.tf_export import tf_export
