+++ tensorflow/tools/api/golden/tensorflow.train.pbtxt
@@ -225,6 +225,10 @@ tf_module {
     mtype: "<type \'type\'>"
   }
   member {
+    name: "VocabInfo"
+    mtype: "<type \'type\'>"
+  }
+  member {
     name: "WorkerSessionCreator"
     mtype: "<type \'type\'>"
   }
@@ -435,6 +439,10 @@ tf_module {
   member_method {
     name: "update_checkpoint_state"
     argspec: "args=[\'save_dir\', \'model_checkpoint_path\', \'all_model_checkpoint_paths\', \'latest_filename\'], varargs=None, keywords=None, defaults=[\'None\', \'None\'], "
+  }
+  member_method {
+    name: "warm_start"
+    argspec: "args=[\'ckpt_to_initialize_from\', \'vars_to_warm_start\', \'var_name_to_vocab_info\', \'var_name_to_prev_var_name\'], varargs=None, keywords=None, defaults=[\'.*\', \'None\', \'None\'], "
   }
   member_method {
     name: "write_graph"
